+++ server/Src/ImageServer/GEOMETRY/GeometricCircle.h
@@ -5,10 +5,10 @@ class GeometricCircle
 {
 public:
 	GeometricCircle();
-	GeometricCircle(const cv::Point& center, int radius);
+	GeometricCircle(const cv::Point& center, float radius);
 	cv::Point centerAsCvPoint() const;
 	int radius() const;
 private:
-	int m_radius;
+	float m_radius;
 	cv::Point m_point;
 };
